fix return type of 'Calendar_GetDaysSinceStartFromDate'

+++ include/g_data.h
@@ -358,7 +358,7 @@ u16 Player_GetActiveSocialLink();
 u8 Player_GetSocialLinkLevel(u16 socialLink);
 PersonaData* Player_GetPersonaByCompendiumIdx(u32 idx);
 
-u8 Calendar_GetDaysSinceStartFromDate(u32 month, u32 day);
+u32 Calendar_GetDaysSinceStartFromDate(u32 month, u32 day);
 u8 Calendar_IsDateInRange(u32 startMonth, u32 startDay, u32 endMonth, u32 endDay);
 u8 Calendar_IsDateInRangeFromDate(u32 monthToTest, u32 dayToTest, u32 startMonth, u32 startDay, u32 endMonth, u32 endDay);
 u8 Calendar_IsDateInRangeFromStart(u32 month, u32 day, u32 range);
